Reduce shared channel usage
