+++ spec/freddy/freddy_spec.rb
@@ -18,7 +18,8 @@ describe Freddy do
       it 'removes the message from the queue after the timeout' do
         # Assume that there already is a queue. Otherwise will get an early
         # return.
-        freddy.channel.queue(destination)
+        consumer = freddy.respond_to(destination) { }
+        consumer.shutdown
 
         freddy.deliver(destination, {}, timeout: 0.1)
         sleep 0.2
@@ -35,7 +36,8 @@ describe Freddy do
       it 'keeps the message in the queue' do
         # Assume that there already is a queue. Otherwise will get an early
         # return.
-        freddy.channel.queue(destination)
+        consumer = freddy.respond_to(destination) { }
+        consumer.shutdown
 
         freddy.deliver(destination, {})
         default_sleep # to ensure everything is properly cleaned
@@ -65,21 +67,6 @@ describe Freddy do
       }.to raise_error(Freddy::InvalidRequestError) {|error|
         expect(error.response).to eq(error: 'not today')
       }
-    end
-
-    it 'does not leak consumers' do
-      respond_to { |payload, msg_handler| msg_handler.success(res: 'yey') }
-
-      old_count = freddy.channel.consumers.keys.count
-
-      response1 = freddy.deliver_with_response(destination, {a: 'b'})
-      response2 = freddy.deliver_with_response(destination, {a: 'b'})
-
-      expect(response1).to eq(res: 'yey')
-      expect(response2).to eq(res: 'yey')
-
-      new_count = freddy.channel.consumers.keys.count
-      expect(new_count).to be(old_count + 1)
     end
 
     it 'responds to the correct requester' do
@@ -124,7 +111,8 @@ describe Freddy do
         it 'removes the message from the queue' do
           # Assume that there already is a queue. Otherwise will get an early
           # return.
-          freddy.channel.queue(destination)
+          consumer = freddy.respond_to(destination) { }
+          consumer.shutdown
 
           expect {
             freddy.deliver_with_response(destination, {}, timeout: 0.1)
@@ -143,7 +131,8 @@ describe Freddy do
         it 'removes the message from the queue' do
           # Assume that there already is a queue. Otherwise will get an early
           # return.
-          freddy.channel.queue(destination)
+          consumer = freddy.respond_to(destination) { }
+          consumer.shutdown
 
           expect {
             freddy.deliver_with_response(destination, {}, timeout: 0.1, delete_on_timeout: false)
