+++ spec/integration/logging_spec.rb
@@ -28,7 +28,6 @@ describe 'Logging' do
 
   it 'logs all produced messages' do
     expect(logger2).to have_received(:debug).with(/Consuming messages on \S+/)
-    expect(logger2).to have_received(:debug).with(/Publishing request to \S+, waiting for response on amq.gen-\S+ with correlation_id .*/)
-    expect(logger2).to have_received(:debug).with(/Producing message {:pay=>"load"} to \S+/)
+    expect(logger2).to have_received(:debug).with(/Publishing request with payload {:pay=>"load"} to \S+, waiting for response on amq.gen-\S+ with correlation_id .*/)
   end
 end
